requirements.txt
openai
python-dotenv
vosk
gpiozero
luma.oled
Pillow
RPi.GPIO
spidev